streamlit_prototype/requirements.txt
streamlit
google-generativeai
websockets
